Add files via upload
Sales analysis tool is developed using Pandas, Matplotlib, Plotly and Streamlit libraries in  python language. This tool performs an analysis of a sales dataset, which should contain columns such as 'Date', 'Amount', 'Status', and 'Category'. Upon analyzing the dataset, the code generates graphical visualizations to provide insights and interpretations of the data.